+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 3.6.1
+- More accurate artifact mainstats.
 # 3.6.0
 **This version includes Breaking Changes**
 - Changed type of User#profilePictureCharacter to `CostumedCharacter | null`.